Move pagination components to subdirectory and migrate them to TypeScript
