+++ packages/ra-ui-materialui/src/list/ListView.tsx
@@ -15,7 +15,7 @@ import {
 
 import Title, { TitlePropType } from '../layout/Title';
 import ListToolbar from './ListToolbar';
-import DefaultPagination from './Pagination';
+import DefaultPagination from './pagination/Pagination';
 import BulkDeleteButton from '../button/BulkDeleteButton';
 import BulkActionsToolbar from './BulkActionsToolbar';
 import DefaultActions from './ListActions';
