+++ packages/ra-ui-materialui/src/list/index.js
@@ -5,9 +5,6 @@ import ListActions from './ListActions';
 import ListGuesser from './ListGuesser';
 import ListToolbar from './ListToolbar';
 import ListView from './ListView';
-import Pagination from './Pagination';
-import PaginationActions from './PaginationActions';
-import PaginationLimit from './PaginationLimit';
 import Placeholder from './Placeholder';
 import SimpleList from './SimpleList';
 import SimpleListLoading from './SimpleListLoading';
@@ -15,6 +12,7 @@ import SingleFieldList from './SingleFieldList';
 
 export * from './filter';
 export * from './datagrid';
+export * from './pagination';
 
 export {
     BulkActionsToolbar,
@@ -24,9 +22,6 @@ export {
     ListActions,
     ListGuesser,
     ListToolbar,
-    Pagination,
-    PaginationActions,
-    PaginationLimit,
     Placeholder,
     SimpleList,
     SimpleListLoading,
