+++ packages/ra-ui-materialui/src/list/pagination/index.ts
@@ -1,0 +1,5 @@
+import Pagination from './Pagination';
+import PaginationActions from './PaginationActions';
+import PaginationLimit from './PaginationLimit';
+
+export { Pagination, PaginationActions, PaginationLimit };
